+++ org.eclipse.jdt.text.tests/src/org/eclipse/jdt/text/tests/performance/TextPerformanceTestCase.java
@@ -459,7 +459,7 @@ public class TextPerformanceTestCase extends TestCase {
 	private static int intValueOf(String stringValue, int defaultValue) {
 		try {
 			if (stringValue != null)
-				return Integer.valueOf(stringValue).intValue();
+				return Integer.valueOf(stringValue);
 		} catch (NumberFormatException e) {
 			// use default
 		}
