+++ org.eclipse.jdt.ui/ui refactoring/org/eclipse/jdt/internal/ui/refactoring/PullUpMemberPage.java
@@ -112,7 +112,7 @@ public class PullUpMemberPage extends UserInputWizardPage {
 		public void modify(final Object element, final String property, final Object value) {
 			if (!ACTION_PROPERTY.equals(property))
 				return;
-			final int action= ((Integer) value).intValue();
+			final int action= ((Integer) value);
 			MemberActionInfo info;
 			if (element instanceof Item) {
 				info= (MemberActionInfo) ((Item) element).getData();
@@ -291,7 +291,7 @@ public class PullUpMemberPage extends UserInputWizardPage {
 	}
 
 	private static void putToStringMapping(final Map<String, Integer> result, final String[] actionLabels, final int actionIndex) {
-		result.put(actionLabels[actionIndex], Integer.valueOf(actionIndex));
+		result.put(actionLabels[actionIndex], actionIndex);
 	}
 
 	private static void setActionForInfos(final MemberActionInfo[] infos, final int action) {
@@ -712,7 +712,7 @@ public class PullUpMemberPage extends UserInputWizardPage {
 			dialog.setBlockOnOpen(true);
 			if (dialog.open() == Window.CANCEL)
 				return;
-			final int action= stringMapping.get(dialog.getSelectedString()).intValue();
+			final int action= stringMapping.get(dialog.getSelectedString());
 			setActionForInfos(selectedMembers, action);
 		} finally {
 			updateWizardPage(preserved, true);
@@ -769,7 +769,7 @@ public class PullUpMemberPage extends UserInputWizardPage {
 		if (commonActionCode == -1)
 			return 0;
 		for (String key : stringMapping.keySet()) {
-			final int action= stringMapping.get(key).intValue();
+			final int action= stringMapping.get(key);
 			if (commonActionCode == action) {
 				for (int i= 0; i < keys.length; i++) {
 					if (key.equals(keys[i]))
